+++ mettle/src/Makefile.am
@@ -52,10 +52,12 @@ libmettle_la_SOURCES += stdapi/ui/osx_desktop.m
 libmettle_la_SOURCES += stdapi/clipboard/osx_clipboard.m
 libmettle_la_SOURCES += stdapi/audio/apple_mic.m
 libmettle_la_SOURCES += stdapi/audio/apple_output.m
+libmettle_la_SOURCES += stdapi/audio/posix_output.c
 endif
 if HOST_LINUX
 libmettle_la_SOURCES += stdapi/webcam/linux_webcam.c
 libmettle_la_SOURCES += stdapi/audio/linux_output.c
+libmettle_la_SOURCES += stdapi/audio/posix_output.c
 endif
 
 libmettle_la_SOURCES += compat/progname.c
